refactor: repath for request component
