+++ src/components/lesseeHome-page/RequestBtn.tsx
@@ -8,7 +8,7 @@ import {
   DialogTitle,
   DialogTrigger,
 } from '@/components/ui/dialog';
-import { RequestForm } from '@/components/request/RequestForm';
+import { RequestForm } from '@/components/lesseeHome-page/RequestForm';
 
 export function RequestBtn() {
   return (
